refactor the source a bit
