+++ src/main.ts
@@ -1,5 +1,5 @@
-import './app.css'
-import App from './App.svelte'
+import './global.css'
+import App from './components/App.svelte'
 
 const app = new App({
   target: document.getElementById('app')!,
